delete GameMng
